+++ src/main/java/com/baichen/rpc/api/AddFallbackImpl.java
@@ -1,0 +1,13 @@
+package com.baichen.rpc.api;
+
+public class AddFallbackImpl implements Add {
+    @Override
+    public int add(int a, int b) {
+        return 0;
+    }
+
+    @Override
+    public int minus(int a, int b) {
+        return 0;
+    }
+}