+++ src/main/java/com/baichen/rpc/breaker/CircuitBreaker.java
@@ -1,6 +1,6 @@
 package com.baichen.rpc.breaker;
 
-import com.baichen.metrics.MetricsData;
+import com.baichen.rpc.metrics.MetricsData;
 
 public interface CircuitBreaker {