+++ src/main/java/com/baichen/rpc/fallback/Fallback.java
@@ -1,0 +1,10 @@
+package com.baichen.rpc.fallback;
+
+import com.baichen.rpc.metrics.MetricsData;
+
+public interface Fallback {
+
+    Object fallback(MetricsData metricsData) throws Exception;
+
+    void recordMetrics(MetricsData metricsData);
+}