+++ src/main/java/com/baichen/rpc/metrics/MetricsData.java
@@ -1,4 +1,4 @@
-package com.baichen.metrics;
+package com.baichen.rpc.metrics;
 
 import com.baichen.rpc.registry.ServiceMateData;
 import lombok.Data;
@@ -15,6 +15,7 @@ public class MetricsData {
 
     private Method method;
     private Object[] args;
+    private Object result;
     private ServiceMateData providerService;
 
     public static MetricsData create(Method method, Object[] args, ServiceMateData providerService) {
@@ -26,9 +27,10 @@ public class MetricsData {
         return metricsData;
     }
 
-    public void complete() {
+    public void complete(Object result) {
         this.success = true;
         this.duration = System.currentTimeMillis() - startTime;
+        this.result = result;
     }
 
     public void completeWithException(Throwable t) {